move duplicated mocking code to a dedicated macro
